Refactoring code. Updating project structure and tests.
