+++ src/components/common/AlertDialog/AlertDialog.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import AlertDialog from '../AlertDialog';
+import AlertDialog from './AlertDialog';
 
 describe("<AlertDialog />", () => {
